+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Fixed
+* IOException is sometimes considered transient
+### Added
+* Retries on metadata fetch
+
 ## [5.0.2] 2023-08-24
 ### Fixed
 * close the http response
