+++ data/src/main/java/com/microsoft/azure/kusto/data/ClientFactory.java
@@ -4,6 +4,7 @@
 package com.microsoft.azure.kusto.data;
 
 import com.microsoft.azure.kusto.data.auth.ConnectionStringBuilder;
+import com.microsoft.azure.kusto.data.http.HttpClientProperties;
 import org.apache.http.impl.client.CloseableHttpClient;
 
 import java.net.URISyntaxException;
