+++ data/src/main/java/com/microsoft/azure/kusto/data/IngestionSourceStorage.java
@@ -2,6 +2,7 @@ package com.microsoft.azure.kusto.data;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.microsoft.azure.kusto.data.http.HttpPostUtils;
 
 class IngestionSourceStorage {
     public String sourceUri;
@@ -11,7 +12,7 @@ class IngestionSourceStorage {
     }
 
     public String toString() {
-        ObjectMapper objectMapper = Utils.getObjectMapper();
+        ObjectMapper objectMapper = HttpPostUtils.getObjectMapper();
         try {
             return objectMapper.writeValueAsString(this);
         } catch (JsonProcessingException e) {
