+++ data/src/main/java/com/microsoft/azure/kusto/data/KustoOperationResult.java
@@ -9,6 +9,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ArrayNode;
 import com.microsoft.azure.kusto.data.exceptions.JsonPropertyMissingException;
 import com.microsoft.azure.kusto.data.exceptions.KustoServiceQueryError;
+import com.microsoft.azure.kusto.data.http.HttpPostUtils;
 import com.microsoft.azure.kusto.data.instrumentation.MonitoredActivity;
 import com.microsoft.azure.kusto.data.instrumentation.SupplierOneException;
 import org.slf4j.Logger;
@@ -36,7 +37,7 @@ public class KustoOperationResult implements Iterator<KustoResultSetTable> {
     private final List<KustoResultSetTable> resultTables = new ArrayList<>();
     private final Iterator<KustoResultSetTable> it;
 
-    private final ObjectMapper objectMapper = Utils.getObjectMapper();
+    private final ObjectMapper objectMapper = HttpPostUtils.getObjectMapper();
 
     public KustoOperationResult(String response, String version) throws KustoServiceQueryError {
         MonitoredActivity.invoke((SupplierOneException<Void, KustoServiceQueryError>) () -> {
