+++ data/src/main/java/com/microsoft/azure/kusto/data/auth/endpoints/WellKnownKustoEndpointsData.java
@@ -2,7 +2,7 @@ package com.microsoft.azure.kusto.data.auth.endpoints;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.microsoft.azure.kusto.data.Utils;
+import com.microsoft.azure.kusto.data.http.HttpPostUtils;
 
 import java.io.InputStream;
 import java.util.ArrayList;
@@ -38,7 +38,7 @@ public class WellKnownKustoEndpointsData {
     private static WellKnownKustoEndpointsData readInstance() {
         try {
             // Beautiful !
-            ObjectMapper objectMapper = Utils.getObjectMapper();
+            ObjectMapper objectMapper = HttpPostUtils.getObjectMapper();
             try (InputStream resourceAsStream = WellKnownKustoEndpointsData.class.getResourceAsStream(
                     "/WellKnownKustoEndpoints.json")) {
                 return objectMapper.readValue(resourceAsStream, WellKnownKustoEndpointsData.class);
