+++ data/src/main/java/com/microsoft/azure/kusto/data/http/HttpClientFactory.java
@@ -1,4 +1,4 @@
-package com.microsoft.azure.kusto.data;
+package com.microsoft.azure.kusto.data.http;
 
 import org.apache.http.HeaderElement;
 import org.apache.http.HeaderElementIterator;
@@ -42,7 +42,16 @@ public class HttpClientFactory {
 
         if (properties.isKeepAlive()) {
             final ConnectionKeepAliveStrategy keepAliveStrategy = new CustomConnectionKeepAliveStrategy(properties.maxKeepAliveTime());
-
+            final ConnectionKeepAliveStrategy keepAliveStrategy2 = (response, context) -> {
+                String header = response.getFirstHeader("Keep-Alive").getValue();
+                if (header == null) {
+                    header = response.getFirstHeader("Connection").getValue();
+                }
+                if (header != null && header.equalsIgnoreCase("keep-alive")) {
+                    return 30 * 1000; // keep the connection alive for 30 seconds
+                }
+                return -1; // close the connection otherwise
+            };
             httpClientBuilder.setKeepAliveStrategy(keepAliveStrategy);
         }
