+++ data/src/main/java/com/microsoft/azure/kusto/data/http/HttpClientProperties.java
@@ -1,4 +1,4 @@
-package com.microsoft.azure.kusto.data;
+package com.microsoft.azure.kusto.data.http;
 
 import org.apache.http.HttpHost;
 
@@ -6,7 +6,6 @@ import org.apache.http.HttpHost;
  * HTTP client properties.
  */
 public class HttpClientProperties {
-
     private final Integer maxIdleTime;
     private final boolean keepAlive;
     private final Integer maxKeepAliveTime;
@@ -51,7 +50,7 @@ public class HttpClientProperties {
      * {@code Keep-Alive} header, the client will use a default keep-alive period indicated by
      * {@linkplain #maxKeepAliveTime()}.
      *
-     * @return whether or not a custom connection keep-alive strategy should be used
+     * @return whether a custom connection keep-alive strategy should be used
      *
      * @see #maxKeepAliveTime()
      * @see <a href="https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Keep-Alive">Keep-Alive</a>
