+++ data/src/test/java/com/microsoft/azure/kusto/data/HttpClientFactoryTest.java
@@ -1,5 +1,7 @@
 package com.microsoft.azure.kusto.data;
 
+import com.microsoft.azure.kusto.data.http.HttpClientFactory;
+import com.microsoft.azure.kusto.data.http.HttpClientProperties;
 import org.apache.http.impl.client.CloseableHttpClient;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
