+++ data/src/test/java/com/microsoft/azure/kusto/data/UtilitiesTest.java
@@ -3,6 +3,7 @@ package com.microsoft.azure.kusto.data;
 import com.microsoft.azure.kusto.data.exceptions.DataServiceException;
 import com.microsoft.azure.kusto.data.exceptions.DataWebException;
 
+import com.microsoft.azure.kusto.data.http.HttpPostUtils;
 import org.apache.http.ProtocolVersion;
 import org.apache.http.message.BasicHttpResponse;
 import org.apache.http.message.BasicStatusLine;
@@ -12,6 +13,9 @@ import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import java.io.IOException;
+import java.net.ConnectException;
+import java.net.UnknownHostException;
 import java.nio.file.Paths;
 import java.util.Objects;
 import java.util.concurrent.TimeUnit;
@@ -45,7 +49,7 @@ class UtilitiesTest {
     @Test
     @DisplayName("Test exception creation when the web response is null")
     void createExceptionFromResponseNoResponse() {
-        DataServiceException error = Utils.createExceptionFromResponse("https://sample.kusto.windows.net", null, new Exception(), "error");
+        DataServiceException error = HttpPostUtils.createExceptionFromResponse("https://sample.kusto.windows.net", null, new Exception(), "error");
         Assertions.assertEquals("POST failed to send request", error.getMessage());
         Assertions.assertFalse(error.isPermanent());
     }
@@ -54,7 +58,7 @@ class UtilitiesTest {
     @DisplayName("Test exception creation on a 404 error")
     void createExceptionFromResponse404Error() {
         BasicHttpResponse basicHttpResponse = getBasicHttpResponse(404);
-        DataServiceException error = Utils.createExceptionFromResponse("https://sample.kusto.windows.net", basicHttpResponse, new Exception(), "error");
+        DataServiceException error = HttpPostUtils.createExceptionFromResponse("https://sample.kusto.windows.net", basicHttpResponse, new Exception(), "error");
         Assertions.assertTrue(error.getStatusCode() != null && error.getStatusCode() == HttpStatus.SC_NOT_FOUND);
     }
 
@@ -84,7 +88,7 @@ class UtilitiesTest {
                 "                \"@permanent\": true\n" +
                 "            }}";
         BasicHttpResponse basicHttpResponse = getBasicHttpResponse(401);
-        DataServiceException error = Utils.createExceptionFromResponse("https://sample.kusto.windows.net", basicHttpResponse, new Exception(), OneApiError);
+        DataServiceException error = HttpPostUtils.createExceptionFromResponse("https://sample.kusto.windows.net", basicHttpResponse, new Exception(), OneApiError);
         Assertions.assertEquals("Query execution has exceeded the allowed limits (80DA0003): ., ActivityId='1234'", error.getMessage());
         Assertions.assertTrue(error.getCause() instanceof DataWebException);
         Assertions.assertTrue(error.isPermanent());
@@ -96,7 +100,7 @@ class UtilitiesTest {
     void createExceptionFromMessageError() {
         String errorMessage = "{\"message\": \"Test Error Message\"}";
         BasicHttpResponse basicHttpResponse = getBasicHttpResponse(401);
-        DataServiceException error = Utils.createExceptionFromResponse("https://sample.kusto.windows.net", basicHttpResponse, new Exception(), errorMessage);
+        DataServiceException error = HttpPostUtils.createExceptionFromResponse("https://sample.kusto.windows.net", basicHttpResponse, new Exception(), errorMessage);
         Assertions.assertEquals("Test Error Message, ActivityId='1234'", error.getMessage());
         Assertions.assertFalse(error.isPermanent());
         Assertions.assertEquals(401, Objects.requireNonNull(error.getStatusCode()).intValue());
@@ -107,7 +111,7 @@ class UtilitiesTest {
     void createExceptionFromBadJson() {
         String errorMessage = "\"message\": \"Test Error Message\"";
         BasicHttpResponse basicHttpResponse = getBasicHttpResponse(401);
-        DataServiceException error = Utils.createExceptionFromResponse("https://sample.kusto.windows.net", basicHttpResponse, new Exception(), errorMessage);
+        DataServiceException error = HttpPostUtils.createExceptionFromResponse("https://sample.kusto.windows.net", basicHttpResponse, new Exception(), errorMessage);
         Assertions.assertEquals("\"message\": \"Test Error Message\", ActivityId='1234'", error.getMessage());
         Assertions.assertFalse(error.isPermanent());
         Assertions.assertEquals(401, Objects.requireNonNull(error.getStatusCode()).intValue());
@@ -118,7 +122,7 @@ class UtilitiesTest {
     void createExceptionFromOtherJson() {
         String errorMessage = "{\"response\": \"Test Error Message\"}";
         BasicHttpResponse basicHttpResponse = getBasicHttpResponse(401);
-        DataServiceException error = Utils.createExceptionFromResponse("https://sample.kusto.windows.net", basicHttpResponse, new Exception(), errorMessage);
+        DataServiceException error = HttpPostUtils.createExceptionFromResponse("https://sample.kusto.windows.net", basicHttpResponse, new Exception(), errorMessage);
         Assertions.assertEquals("{\"response\": \"Test Error Message\"}, ActivityId='1234'", error.getMessage());
         Assertions.assertFalse(error.isPermanent());
         Assertions.assertEquals(401, Objects.requireNonNull(error.getStatusCode()).intValue());
@@ -129,7 +133,7 @@ class UtilitiesTest {
     void createExceptionFromBlankErrorMessage() {
         String errorMessage = " ";
         BasicHttpResponse basicHttpResponse = getBasicHttpResponse(401);
-        DataServiceException error = Utils.createExceptionFromResponse("https://sample.kusto.windows.net", basicHttpResponse, new Exception(), errorMessage);
+        DataServiceException error = HttpPostUtils.createExceptionFromResponse("https://sample.kusto.windows.net", basicHttpResponse, new Exception(), errorMessage);
         Assertions.assertEquals("Http StatusCode='http/1.1 401 Some Error', ActivityId='1234'", error.getMessage());
         Assertions.assertFalse(error.isPermanent());
         Assertions.assertEquals(401, Objects.requireNonNull(error.getStatusCode()).intValue());
@@ -146,6 +150,19 @@ class UtilitiesTest {
     void extractFileNameFromCommandLine() {
         String cmdLine = Paths.get(" home", "user", "someFile.jar").toString() + " -arg1 val";
         Assertions.assertEquals(UriUtils.stripFileNameFromCommandLine(cmdLine), "someFile.jar");
+    }
+
+    @Test
+    @DisplayName("Test exception creation from a message object")
+    void isRetrieable() {
+        IOException e = new UnknownHostException("Doesnt exist");
+        Assertions.assertFalse(Utils.isRetriableIOException(e));
+
+        e = new org.apache.http.conn.HttpHostConnectException(new ConnectException("Connection refused"),null);
+        Assertions.assertFalse(Utils.isRetriableIOException(e));
+
+        e = new org.apache.http.conn.HttpHostConnectException(new ConnectException("Connection timed out"),null);
+        Assertions.assertTrue(Utils.isRetriableIOException(e));
     }
 
     @NotNull
