+++ quickstart/src/main/java/com/microsoft/azure/kusto/quickstart/SampleApp.java
@@ -7,6 +7,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.microsoft.azure.kusto.data.Client;
 import com.microsoft.azure.kusto.data.ClientFactory;
 import com.microsoft.azure.kusto.data.StringUtils;
+import com.microsoft.azure.kusto.data.http.HttpPostUtils;
 import com.microsoft.azure.kusto.data.instrumentation.MonitoredActivity;
 import com.microsoft.azure.kusto.data.instrumentation.Tracer;
 import com.microsoft.azure.kusto.ingest.IngestClient;
@@ -345,7 +346,7 @@ public class SampleApp {
     private static ConfigJson loadConfigs() {
         File configFile = new File(".\\" + SampleApp.configFileName);
         try {
-            ObjectMapper mapper = com.microsoft.azure.kusto.data.Utils.getObjectMapper();
+            ObjectMapper mapper = HttpPostUtils.getObjectMapper();
             mapper.configure(MapperFeature.ACCEPT_CASE_INSENSITIVE_ENUMS, true);
             return mapper.readValue(configFile, ConfigJson.class);
