+++ samples/src/main/java/TableStatus.java
@@ -3,7 +3,7 @@
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
-import com.microsoft.azure.kusto.data.Utils;
+import com.microsoft.azure.kusto.data.http.HttpPostUtils;
 import com.microsoft.azure.kusto.data.auth.ConnectionStringBuilder;
 import com.microsoft.azure.kusto.ingest.IngestClient;
 import com.microsoft.azure.kusto.ingest.IngestClientFactory;
@@ -47,7 +47,7 @@ public class TableStatus {
                 statuses = ingestionResult.getIngestionStatusCollection();
             }
 
-            ObjectMapper objectMapper = Utils.getObjectMapper();
+            ObjectMapper objectMapper = HttpPostUtils.getObjectMapper();
             JavaTimeModule module = new JavaTimeModule();
             module.addSerializer(Instant.class, new InstantSerializerWithMilliSecondPrecision());
             objectMapper.registerModule(module);
